add: UserBlock component
