+++ project/src/pages/authorization-screen/authorization-screen.tsx
@@ -1,6 +1,6 @@
 import SignInForm from '../../components/sign-in-form/sign-in-form';
 import Footer from '../../components/footer/footer';
-import UnauthorizedUserHeader from '../../components/unauthorized-user-header/unauthorized-user-header';
+import UnauthorizedUserHeader from '../../components/user-header/unauthorized-user-header';
 
 function AuthorizationScreen(): JSX.Element {
   return (
